% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomass}
\alias{biomass}
\title{Species biomasses in communities X and Y}
\format{A data frame with 30 rows and 3 variables:
\describe{
  \item{species}{list of species ids}
  \item{biomassX}{mass of species in community X, in grams/m^2}
  \item{biomassY}{mass of species in community Y, in grams/m^2}
}}
\source{
Placeholder
}
\usage{
biomass
}
\description{
A small example dataset containing the biomass and names of 30 species in two plant communities.
}
\keyword{datasets}

