% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braycurtis_functions.R
\name{braycurtis.single}
\alias{braycurtis.single}
\title{Low-level wrapper function for calculating Bray-Curtis index similarity between a pair of communities}
\usage{
braycurtis.single(sps, func, commX)
}
\arguments{
\item{sps}{A vector of species' names}

\item{func}{A numerical vector of species' functions}

\item{commX}{A reference community}
}
\value{
This function returns a matrix with a single row, and a column with the Bray-Curtis index.
}
\description{
Given a list of species names and their functions, and a reference community,
calculate the Bray-Curtis index and return it. This is a  low-level function used inside of 
higher-level functions (ie, \code{pairwise.braycurtis()}) that automate the pairwise comparison 
of many communities. Note that if a comparison fails, NA is returned instead. So far this only 
seems to happen when comparisons are made between two communities each containing the same, 
single species.
}
\examples{

# Generate mock community data:
set.seed(36)
cm1<-data.frame(sps=LETTERS[seq(1,6)],func=rpois(6,lambda = 2))
cm2<-data.frame(sps=LETTERS[seq(1,6)],func=rpois(6,lambda = 2))

# Compare community 2's species and function lists to community 1
braycurtis.single(sps=cm2$sps,func=cm2$func,commX=cm1)

}

