% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cedarcreek}
\alias{cedarcreek}
\title{Plant biomass data from Cedar Creek, MN}
\format{A data frame with 30 rows and 3 variables:
\describe{
  \item{Plot}{Plot ID}
  \item{NTrt}{Nitrogen treatment code}
  \item{NAdd}{Amount of nitrogen added}
  \item{Species}{Species name}
  \item{Biomass}{Biomass of species in a given community, in grams/m^2}
}}
\source{
Tilman and Downing 1997, and others.
}
\usage{
cedarcreek
}
\description{
Data on species' biomass from Cedar Creek experiment e001, Field D, 1992. This experiment contains
a nitrogen addition manipulation, with a control (no addition) and seven levels of increasing nitrogen
addition.
}
\keyword{datasets}

