% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementarity_and_selection.R
\name{comp.sel}
\alias{comp.sel}
\title{Calculate estimates of Complementarity and Selection}
\usage{
comp.sel(data, even.seeding = FALSE)
}
\arguments{
\item{data}{This is a data frame containing columns 'mono' (species' monoculture abundance), 'mixed' (species polyculture abundance), and 'init.mixed' (initial relative abundance of species in polyculture)}

\item{even.seeding}{If no 'init.mixed' column is provided, and even.seeding is set to TRUE, equal relative abundances are inferred within the function}
}
\value{
CE the complementarity effect

SE the selection effect
}
\description{
Calculate the complementarity and selection effects (Loreau & Hector 2001) given data on
the abundance of species grown in mono- and poly-culture.
}
\examples{

Format 1:
set.seed(1)
d1<-data.frame(sps=c('A','B','C','D','E'),mono=1+runif(5),init.mixed=1/5,mixed=1+runif(5,min=0.2))

comp.sel(d1)

# Format 2:
set.seed(1)
d2<-data.frame(sps=c('A','B','C','D','E'),mono=1+runif(5),mixed=1+runif(5,min=0.2))

comp.sel(d2,even.seeding = T)

}

