% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementarity_and_selection.R
\name{cov2}
\alias{cov2}
\title{Calculate covariance}
\usage{
cov2(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{Another vector}
}
\value{
Covariance between x and y
}
\description{
Calculate covariance
}
\examples{
x<-runif(10)
y<-runif(10)
cov2(x,y)

}

