% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_FUNCTIONS_050916.R
\name{data.setup}
\alias{data.setup}
\title{Data setup for Price Equation calculations}
\usage{
data.setup(input)
}
\arguments{
\item{input}{A list of species names}
}
\value{
comm  A data frame containing columns:
}
\description{
Take two separate community lists and create a single data set, which is ready
for calculating the Price Equation partitions.
}

