% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_summary_stats.R
\name{delta.vars}
\alias{delta.vars}
\title{Calculate the difference in variances between distributions.}
\usage{
delta.vars(x, y)
}
\arguments{
\item{x}{A categorical value (factor) with two levels.}

\item{y}{A list of response values.}
}
\value{
The difference in variances.
}
\description{
Calculate the difference in variances between distributions.
}
\examples{

y <- c(rnorm(20,mean=10,sd=2),rnorm(20,mean=25,sd=6))
x <- factor(c(rep('A',20),rep('B',20)))

delta.vars(x,y)

}

