% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_FUNCTIONS.R
\name{dist.mat.size}
\alias{dist.mat.size}
\title{Rough estimate of memory requirements of distance matrix}
\usage{
dist.mat.size(n)
}
\arguments{
\item{n}{This is the number of rows in the distance matrix}
}
\value{
Estimated memory requirement in GB
}
\description{
Calculate the amount of memory needed to store a distance matrix, before calculating the
distance matrix. This is a prudent way to prevent running out of RAM when calculating a
distance matrix. (Code implemented from response on R help list).
}
\examples{
dist.mat.size(200)

}

