% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_summary_stats.R
\name{dist.test}
\alias{dist.test}
\title{Test the difference in means between distributions.}
\usage{
dist.test(x, y)
}
\arguments{
\item{x}{A categorical value (factor) with two levels}

\item{y}{A list of response values}
}
\value{
A p-value coming from an anova/t-test
}
\description{
This function computes a simple t-test, determining a p-value for the difference in means
between two distributions, assuming homogeneity of variance.
}
\examples{

y <- c(rnorm(20,mean=10,sd=2),rnorm(20,mean=25,sd=6))
x <- factor(c(rep('A',20),rep('B',20)))

dist.test(x,y)

}

