% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_FUNCTIONS_050916.R
\name{expit}
\alias{expit}
\title{Logistic link function}
\usage{
expit(x)
}
\arguments{
\item{x}{A number}
}
\value{
Transformed values on the probability scale
}
\description{
Logistic link function
}
\examples{
expit(0)
expit(1)
}

