% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_FUNCTIONS.R
\name{group.columns}
\alias{group.columns}
\title{Merge pairs of grouping columns}
\usage{
group.columns(x, gps, drop = F)
}
\arguments{
\item{x}{A data frame}

\item{gps}{A list of grouping column(s)}

\item{drop}{Drop original, unpaired columns after grouping? TRUE/FALSE}
}
\value{
This function returns a data frame
}
\description{
Take a pair of columns, distinguished with suffix .x and .y, and merge them into a
single column without suffixes. This is helpful for post-processing the output of
pairwise price calculations.
}
\examples{

# write example

}

