% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_functions.R
\name{jaccard.column}
\alias{jaccard.column}
\title{Wrapper function for calculating Jaccard index for a list of communities}
\usage{
jaccard.column(sps, func, dat)
}
\arguments{
\item{sps}{A vector of species' names for the reference community}

\item{func}{A numerical vector of species' ecosystem functions in the reference
community}

\item{dat}{A grouped data frame of species' names and ecosystem functions, which 
must contain at least one grouping variable, as created by dplyr.}
}
\value{
This function returns a data set of Jaccard indices for each community 
(defined by the grouping variables) compared against the reference community
}
\description{
Given a list of species names and their ecosystem functions, this function generates
a reference community, and then compares the reference community against a set of
other communities (including species and their ecosystem function) supplied in a
separate, grouped data frame. This is a low-level function that invokes 
\code{jaccard.single()} and is called by higher-level functions such as 
\code{pairwise.jaccard()}, which automates the pairwise comparison of many communities.
}
\examples{

# write example

}

