% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_functions.R
\name{jaccard.single}
\alias{jaccard.single}
\title{Low-level wrapper function for calculating Jaccard index similarity between a pair of communities}
\usage{
jaccard.single(sps, func, commX)
}
\arguments{
\item{sps}{A vector of species' names}

\item{func}{A numerical vector of species' functions}

\item{commX}{A reference community}
}
\value{
This function returns a matrix with a single row, and a column with the Jaccard index.
}
\description{
Given a list of species names and their functions, and a reference community,
calculate the Jaccard index and return it. This is a  low-level function used inside of 
higher-level functions (ie, \code{pairwise.jaccard()}) that automate the pairwise comparison 
of many communities. Note that if a comparison fails, NA is returned instead. So far this only 
seems to happen when comparisons are made between two communities each containing the same, 
single species.
}
\examples{

# write example

}

