% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_plotting.R
\name{leap.zig.both}
\alias{leap.zig.bef}
\alias{leap.zig.both}
\alias{leap.zig.cafe}
\alias{leap.zig.price}
\title{Internal functions for generating vector plots of ecosystem function components.}
\usage{
leap.zig.both(tmp, xlim = NA, ylim = NA, loc.standardize = TRUE,
  error.bars = FALSE, raw.points = TRUE, vectors = TRUE,
  group.vars = NULL, legend = TRUE, old.plot = NA, add = FALSE)

leap.zig.bef(tmp, xlim = NA, ylim = NA, loc.standardize = TRUE,
  error.bars = FALSE, raw.points = TRUE, vectors = TRUE,
  group.vars = NULL, legend = TRUE, old.plot = NA, add = FALSE)

leap.zig.cafe(tmp, xlim = NA, ylim = NA, loc.standardize = TRUE,
  error.bars = FALSE, raw.points = TRUE, vectors = TRUE,
  group.vars = NULL, legend = TRUE, old.plot = NA, add = FALSE)

leap.zig.price(tmp, xlim = NA, ylim = NA, loc.standardize = TRUE,
  error.bars = FALSE, raw.points = TRUE, vectors = TRUE,
  group.vars = NULL, legend = TRUE, old.plot = NA, add = FALSE)
}
\arguments{
\item{tmp}{Data to plot}

\item{xlim}{Plot's x limits}

\item{ylim}{Plot's y limits}

\item{error.bars}{Plot error bars}

\item{raw.points}{Plot raw data points at level of community pairs}

\item{vectors}{Plot averaged vectors}

\item{legend}{Show legend}

\item{old.plot}{ggplot object from previous \code{leap.zig()} call}

\item{add}{Add new plot to object provided in old.plot option}

\item{loc.standarize}{Are these standardized vectors}
}
\value{
A ggplot object.
}
\description{
This suite of functions is accessed by \code{\link{leap.zig()}} and produces vector plots 
using BEF, CAFE, and Price components. \code{leap.zig.both} plots both the BEF and CAFE vectors.
}
\section{Functions}{
\itemize{
\item \code{leap.zig.bef}: Plot vectors for BEF components

\item \code{leap.zig.cafe}: Plot vectors for CAFE components

\item \code{leap.zig.price}: Plot vectors for Price components
}}
\examples{

# Load data and run pairwise comparisons  of communities
cc2<-group_by(cedarcreek,NTrt,NAdd,Plot)
pp<-pairwise.price(cc2,species='Species',func='Biomass')

# Organize/format the results, and pull out a subset using NTrt=1 as the control site
pp<-group.columns(pp,gps=c('NTrt','NAdd'))
pp<-pp[pp$NTrt.x==1,]
dat1<-pp[pp$NAdd \%in\% c('0 27.2'),]

# Process and plot the result
tmp <- process.data.bef(dat1, group.vars=NULL, standardize=F)
leap.zig.bef(tmp, loc.standardize=F, group.vars=NULL,raw.points = F,legend = F)

}

