% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_plotting.R
\name{leap.zig}
\alias{leap.zig}
\title{Plot changes in ecosystem function between communities as vector diagram.}
\usage{
leap.zig(data, type = "cafe", group.vars = NULL, standardize = TRUE, ...)
}
\arguments{
\item{data}{Pairwise Price data}

\item{type}{Specify the type of vector diagram to draw ("cafe","bef","both", or "price")}

\item{group.vars}{A vector of grouping variables, if any}

\item{standardize}{Should ecosystem function values be standardized against baseline}

\item{...}{Additional graphical options passed to lower-level functions. See \code{\link{leap.zig.bef}}, \code{\link{leap.zig.cafe}}, \code{\link{leap.zig.both}}, \code{\link{leap.zig.price}}}
}
\value{
A ggplot object.
}
\description{
This wrapper function takes the data generated by \code{\link{pairwise.price()}} and produces vector
plots using either BEF, CAFE, or 5-part Price components.
}
\examples{

# write one

}

