% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_FUNCTIONS_050916.R
\name{lwCImean}
\alias{lwCImean}
\title{Asymptotic 95% confidence interval}
\usage{
lwCImean(x)
}
\arguments{
\item{x}{A list of numeric values}
}
\value{
Confidence interval value
}
\description{
Asymptotic 95% confidence interval
}
\examples{
x<-runif(20)
lwCImean(x)

}

