% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_FUNCTIONS.R
\name{mean.CI}
\alias{mean.CI}
\title{Asymptotic 95% confidence interval}
\usage{
\method{mean}{CI}(x, type)
}
\arguments{
\item{x}{A list of numeric values}

\item{type}{Upper ("upr") or lower ("lwr") CI boundary?}
}
\value{
Confidence interval value
}
\description{
Asymptotic 95% confidence interval
}
\examples{
x<-runif(20)
mean.CI(x,type="lwr")
mean.CI(x,type="upr")

}

