% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_FUNCTIONS.R
\name{meanCI}
\alias{meanCI}
\title{Asymptotic 95 percent confidence interval.}
\usage{
meanCI(x, type)
}
\arguments{
\item{x}{A list of numeric values}

\item{type}{Upper ("upr") or lower ("lwr") CI boundary}
}
\value{
Confidence interval value
}
\description{
This function computes the upper or lower 95 percent CI based on an asymptotic 
approximation and assumption of normality.
}
\examples{

x<-runif(20)

meanCI(x,type="lwr")
meanCI(x,type="upr")

}

