% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_FUNCTIONS.R
\name{pairwise.price}
\alias{pairwise.price}
\title{Calculate the Price equation partition for all possible community pairs}
\usage{
pairwise.price(x, species = "Species", func = "Function")
}
\arguments{
\item{x}{A grouped data set, with grouping variables defined as in dplyr operations}

\item{species}{The name of the column in \code{x} containing species ID's}

\item{func}{The name of the column in \code{x} containing species' ecosystem function}
}
\value{
This function returns a data set of the Price equation components 
 corresponding to pairs of communities, identified by one or more grouping variables,
 which are provided in pairs of columns with the format: groupvar1.x groupvar1.y, etc.
 These can be conveniently re-combined using the \code{group.columns()} command.
}
\description{
Given a grouped data set containing a species ID column and a column of species'
ecosystem functions, this function returns a dataset of Price components that results
from comparing all pairwise combinations of unique communities as defined by the 
grouping variable(s).
}
\examples{

# write example

}

