% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_functions.R
\name{price.part}
\alias{price.part}
\title{Calculate the Price equation partition for two communities}
\usage{
price.part(comm, quiet = F, sps.level = F)
}
\arguments{
\item{comm}{A data frame formatted according to the \code{\link{data.setup()}}
function, or if formatting independently, a data frame with the following columns.}

\item{quiet}{Silence error messages? TRUE/FALSE}

\item{sps.level}{Provide species-level contributions to 5 Price components? TRUE/FALSE}
}
\value{
If \code{sps.level=FALSE}, a data frame of Price equation components.
   \item{SRE.L}{species richness effect (loss of species)}
   \item{SRE.G}{species richness effect (gain of species)}
   \item{SIE.L}{species identity effect (loss of species)}
   \item{SIE.G}{species identity effect (gain of species)}
   \item{CDE}{context dependent effect}
   \item{SL}{sum of SRE.L and SIE.L}
   \item{SG}{sum of SRE.G and SIE.G}
   \item{SR}{sum of SRE.L and SRE.G}
   \item{CE}{sum of SIE.G, SIE.L, and CDE}
   \item{x.func}{Total function in community X}
   \item{y.func}{Total function in community Y}
   \item{x.rich}{Number of species in community X}
   \item{y.rich}{Number of species in community Y}
   \item{c.rich}{Number of shared species between X and Y}

If \code{sps.level=TRUE}, a list containing the above information in the first slot, and a data frame of individual species' contributions to each Price component in the second slot.
}
\description{
Take a (formatted) list of species and functions for two communities, and calculate
the Price equation partition comparing the communities.
}
\details{
Extra thoughts on how the interpret the Price equation partitions, which may get relocated into a vignette.

Comments on SIE.L. If a species x' from x is lost in y, SIE.L will increase if x' 
is less productive on average, and SIE.L will decrease if x' is more productive 
than average. If a species x' from x is NOT lost in y, SIE.L will increase if x' 
is more productive on average, and SIE.L will decrease if x' is less productive 
than average. Overall, high/positive values of SIE.L mean that weak species were 
lost and good species were retained. Noteably, SIE.L will not be affected by new
species that y gains relative to what is shared or lost. Average species have 
little effect on the value of SIE.L. If either barely any or almost all species 
occur in common between communities x and y, then the few species that are kept 
(or lost) will have a particularly large influence on the value of SIE.L. Overall, 
SIE.L will probably be smaller in this case, and greatly affected by whether the
species lost/gained are more or less productive.

Comments on SIE.G. A less productive than average species in y (-1) makes a negative
contribution to SIE.G if it is NOT in community x, and a positive contribution to 
SIE.G if it is in community x. A more productive than average species in y (+1) 
makes a negative contribution to SIE.G if it is in x, and a positive contribution 
to SIE.G if it does NOT occur in community x. A positive SIE.G occurs when less
productive than average members of y also occured in x, and more productive than
average species in y do not occur in x.
}
\examples{

formatted.data<-data.setup(list(biomass))
price.part(formatted.data) 

}

