% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_functions.R
\name{price.part.single}
\alias{price.part.single}
\title{Low-level wrapper function for applying Price partition to a pair of communities}
\usage{
price.part.single(sps, func, commX)
}
\arguments{
\item{sps}{A vector of species' names}

\item{func}{A numerical vector of species' functions}

\item{commX}{A reference or 'baseline' community}
}
\value{
This function returns a matrix with a single row, and columns consisting of 
Price equation components.
}
\description{
Given a list of species names and their functions, and a reference community,
calculate the full set of Price partition components and return them. This is a 
low-level function used inside of higher-level functions (ie, \code{price.part.all()})
that automate the pairwise comparison of many communities.
}
\examples{

# write example

}

