% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_plotting.R
\name{process.data.cafe}
\alias{process.data.bef}
\alias{process.data.cafe}
\alias{process.data.price}
\title{Process data from a single treatment prior to making a vector plot}
\usage{
process.data.cafe(data, group.vars = NULL, standardize = TRUE)

process.data.bef(data, group.vars = NULL, standardize = TRUE)

process.data.price(data, group.vars = NULL, standardize = TRUE)
}
\arguments{
\item{data}{Pairwise Price data}

\item{group.vars}{A vector of grouping variables, if any}

\item{standardize}{Should ecosystem function values be standardized against baseline? T/F}
}
\value{
A list of three data sets with different levels of aggregation used in subsequent vector plots. Although still exported, this function has become essentially an internal function, called by members of the leap.zig family of functions, and may rarely be useful to call directly.
}
\description{
These functions take the data generated by pairwise.price() and processes it into the
terms needed in CAFE, BEF, or Price component vector plots.
}
\section{Functions}{
\itemize{
\item \code{process.data.bef}: Process data for plotting BEF components.

\item \code{process.data.price}: Process data for plotting 5-part Price components.
}}
\examples{

# write one
set.seed(36)

# Data frame containing multiple communities we want to compare
cms<-data.frame(comm.id=sort(rep(seq(1,3),6)),
                species=rep(LETTERS[seq(1,6)],3),
                func=rpois(6*3,lambda = 2))
                
#Identify one (or more) grouping columns
cms<-group_by(cms,comm.id)

# Perform pairwise comparisons of all communities in cms identified by comm.id
pp<-pairwise.price(cms,species='species',func='func')
pp<-group.columns(pp,gps=c('comm.id'))

process.data.cafe(data=pp,group.vars='comm.id')
process.data.bef(data=pp,group.vars='comm.id')
process.data.price(data=pp,group.vars='comm.id')
   
}

