% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_FUNCTIONS_050916.R
\name{se}
\alias{se}
\title{Standard error}
\usage{
se(x)
}
\arguments{
\item{x}{A list of numeric values}
}
\value{
Standard error
}
\description{
Standard error
}
\examples{
x<-runif(29)
se(x)
}

