% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_summary_stats.R
\name{trt.mean}
\alias{trt.mean}
\title{Calculate the mean of a specific treatment's distribution.}
\usage{
trt.mean(x, y, ind)
}
\arguments{
\item{x}{A categorical value (factor) with two levels.}

\item{y}{A list of response values.}

\item{ind}{The index of the specific treatment to return a mean for.}
}
\value{
Treatment mean.
}
\description{
Calculate the mean of a specific treatment's distribution.
}
\examples{

y <- c(rnorm(20,mean=10,sd=2),rnorm(20,mean=25,sd=6))
x <- factor(c(rep('A',20),rep('B',20)))

trt.mean(x,y,1)
trt.mean(x,y,2)

}

