% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_plotting.R
\name{leap.zig.bef}
\alias{leap.zig.bef}
\title{Plot changes in ecosystem function using BEF components.}
\usage{
leap.zig.bef(tmp, xlim = NA, ylim = NA, loc.standardize = TRUE,
  error.bars = FALSE, raw.points = TRUE, vectors = TRUE,
  all.vectors = FALSE, legend = TRUE, old.plot = NA, main = "",
  linetype = 1, add = FALSE)
}
\arguments{
\item{tmp}{Data to plot}

\item{xlim}{Plot's x limits}

\item{ylim}{Plot's y limits}

\item{error.bars}{Plot error bars}

\item{raw.points}{Plot raw data points at level of community pairs}

\item{vectors}{Plot averaged vectors}

\item{all.vectors}{Plot vectors at level of community pairs}

\item{legend}{Show legend}

\item{old.plot}{ggplot object from previous \code{leap.zig()} call}

\item{main}{Plot title}

\item{linetype}{Type of line to draw for vector}

\item{add}{Add new plot to object provided in old.plot option}

\item{loc.standarize}{Are these standardized vectors}
}
\value{
A ggplot object.
}
\description{
This function is called by \code{\link{leap.zig()}} and produces vector plots 
using BEF components, overlaid on top of each other.
}
\examples{

# write several

}

