% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_plotting.R
\name{leap.zig.both}
\alias{leap.zig.both}
\title{Plot changes in ecosystem function using CAFE & BEF components.}
\usage{
leap.zig.both(tmp, xlim = NA, ylim = NA, loc.standardize = TRUE,
  error.bars = FALSE, raw.points = TRUE, vectors = TRUE, legend = TRUE,
  old.plot = NA, linetype = 1, add = FALSE)
}
\arguments{
\item{tmp}{Data to plot}

\item{xlim}{Plot's x limits}

\item{ylim}{Plot's y limits}

\item{error.bars}{Plot error bars}

\item{raw.points}{Plot raw data points at level of community pairs}

\item{vectors}{Plot averaged vectors}

\item{legend}{Show legend}

\item{old.plot}{ggplot object from previous \code{leap.zig()} call}

\item{linetype}{Type of line to draw for vector}

\item{add}{Add new plot to object provided in old.plot option}

\item{loc.standarize}{Are these standardized vectors}
}
\value{
A ggplot object.
}
\description{
This function is called by \code{\link{leap.zig()}} and produces vector plots 
using BEF and CAFE components, overlaid on top of each other.
}
\examples{

# write several

}

