% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_plotting.R
\name{process.data.cafe}
\alias{process.data.bef}
\alias{process.data.cafe}
\alias{process.data.price}
\title{Process single treatment's worth of data for plotting CAFE components.}
\usage{
process.data.cafe(data, group.vars = NULL, standardize = TRUE)

process.data.bef(data, group.vars = NULL, standardize = TRUE)

process.data.price(data, group.vars = NULL, standardize = TRUE)
}
\arguments{
\item{data}{Pairwise Price data}

\item{group.vars}{A vector of grouping variables, if any.}

\item{standardize}{Should ecosystem function values be standardized against baseline}
}
\value{
A list of three data sets with different levels of aggregation used in subsequent vector plots.
}
\description{
This function takes the data generated by pairwise.price() and processes it into 
terms needed in CAFE vector plots.
}
\section{Functions}{
\itemize{
\item \code{process.data.bef}: Process data for plotting BEF components.

\item \code{process.data.price}: Process data for plotting 5-part Price components.
}}
\examples{

# write one

}

