% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_plotting.R
\name{process.data.cafe}
\alias{process.data.cafe}
\title{Process single treatment's worth of data for plotting CAFE components.}
\usage{
process.data.cafe(data, group.vars = NULL, standardize = TRUE)
}
\arguments{
\item{data}{Pairwise Price data}

\item{group.vars}{A vector of grouping variables, if any.}

\item{standardize}{Should ecosystem function values be standardized against baseline}
}
\value{
A list of three data sets with differen levels of aggregation used in subsequent vector plots.
}
\description{
This function takes the data generated by pairwise.price() and processes it into 
terms needed in CAFE vector plots.
}
\examples{

# write one

}

