% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_FUNCTIONS.R
\name{se}
\alias{se}
\title{Standard error.}
\usage{
se(x)
}
\arguments{
\item{x}{A list of numeric values}
}
\value{
Standard error
}
\description{
Short-cut for calculating the standard error of a vector of values
}
\examples{
x<-runif(29)
se(x)
}

