% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_summary_stats.R
\name{test.partitions}
\alias{test.partitions}
\title{Run a set of significance tests on vector plot components.}
\usage{
test.partitions(data, type = "both", treat.var, control, standardize = T,
  print = F, plot = F)
}
\arguments{
\item{data}{}

\item{type}{Specify which kind of components ("bef", "cafe", "both", "price")}

\item{treat.var}{Specify which column in data contains the treatment variable.}

\item{control}{Control variable}

\item{standardize}{Use standardized variables}

\item{print}{Print table}

\item{plot}{Display plot}
}
\value{
Table of significance tests for the means and variances of vector components.
}
\description{
This functions tests the significance of the changes in richness in function associated with the
vectors of a vector plot. It compares a single baseline (or control) community against a comparison
(treatment) community. This can be calculated for the BEF, CAFE, and 5-part Price components.
}
\examples{

# write one.

}

